+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 
 ## [Unreleased]
 
-## Added
+### Added
 
-- GUI window grabs keyboard focus.
+- A preset can be saved via "Manage Preset...". It may replace an existing preset or create a new one. The user also can delete a preset.
 
 ## [v1.0.1] - 2026-06-10